define legacy v1 checkpoints in checkpoint_v1.go
