+++ pkg/kubelet/allocation/state/checkpoint.go
@@ -26,10 +26,6 @@ import (
 
 var _ checkpointmanager.Checkpoint = &Checkpoint{}
 
-type PodResourceCheckpointInfo struct {
-	Entries PodResourceInfoMap `json:"entries,omitempty"`
-}
-
 // Checkpoint represents a structure to store pod resource allocation checkpoint data
 type Checkpoint struct {
 	// Data is a serialized PodResourceAllocationInfo
